v0.1
Yahoo!

I did it! 

It works!!

This is a game where you, the user, can choose to fight either a bug, a fox, a man, or the ogre, Malgroth.

The two of you then fight until one of you dies. 

If you win, you might level up, gain an item, or find other things that will help you eventually defeat the evil ogre, Malgroth!

I intend to continue to develop this as I learn more in the Hack Reactor immersive. Starting January 2021!